Move check_and_create colors
