+++ cellrank/tools/estimators/_cflare.py
@@ -19,7 +19,7 @@ from cellrank.tools._utils import (
     _complex_warning,
     _get_connectivities,
 )
-from cellrank.tools._colors import _convert_to_hex_colors, _create_categorical_colors
+from cellrank.tools._colors import _create_categorical_colors
 from cellrank.tools._lineage import Lineage
 from cellrank.tools._constants import _dp, _probs, _colors, _lin_names
 from cellrank.tools.kernels._kernel import KernelExpression
@@ -654,25 +654,6 @@ class CFLARE(BaseEstimator):
         c = c_ / np.max(c_)
 
         return c
-
-    def _check_and_create_colors(self):
-        n_cats = len(self._meta_states.cat.categories)
-        color_key = _colors(self._rc_key)
-
-        if self._meta_states_colors is None:
-            if color_key in self._adata.uns and n_cats == len(
-                self._adata.uns[color_key]
-            ):
-                logg.debug("Loading colors from `.adata` object")
-                self._meta_states_colors = _convert_to_hex_colors(
-                    self._adata.uns[color_key]
-                )
-            else:
-                self._meta_states_colors = _create_categorical_colors(n_cats)
-                self._adata.uns[color_key] = self._meta_states_colors
-        elif len(self._meta_states_colors) != n_cats:
-            self._meta_states_colors = _create_categorical_colors(n_cats)
-            self._adata.uns[color_key] = self._meta_states_colors
 
     def _get_restriction_to_main(self) -> Tuple[Series, np.ndarray]:
         """
